+++ scripts/build.js
@@ -24,7 +24,12 @@ function build() {
 
   Browserify(inputFile, {standalone: 'baidubce.sdk'})
     .transform('babelify', {
-      presets: ['@babel/preset-env']
+      presets: ['@babel/preset-env'],
+      plugins: [
+        '@babel/plugin-transform-async-to-generator',
+        ['@babel/plugin-syntax-optional-chaining-assign', {version: '2023-07'}],
+        '@babel/plugin-transform-nullish-coalescing-operator'
+      ]
     })
     .bundle()
     .pipe(fs.createWriteStream(outputJS))